+++ quanmiaolightapp-20240801/alibabacloud_quanmiaolightapp20240801/models.py
@@ -4072,6 +4072,422 @@ class RunMarketingInformationWritingResponse(TeaModel):
         return self
 
 
+class RunNetworkContentAuditRequestTags(TeaModel):
+    def __init__(
+        self,
+        tag_define_prompt: str = None,
+        tag_name: str = None,
+    ):
+        self.tag_define_prompt = tag_define_prompt
+        self.tag_name = tag_name
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.tag_define_prompt is not None:
+            result['tagDefinePrompt'] = self.tag_define_prompt
+        if self.tag_name is not None:
+            result['tagName'] = self.tag_name
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('tagDefinePrompt') is not None:
+            self.tag_define_prompt = m.get('tagDefinePrompt')
+        if m.get('tagName') is not None:
+            self.tag_name = m.get('tagName')
+        return self
+
+
+class RunNetworkContentAuditRequest(TeaModel):
+    def __init__(
+        self,
+        business_type: str = None,
+        content: str = None,
+        extra_info: str = None,
+        model_id: str = None,
+        output_format: str = None,
+        tags: List[RunNetworkContentAuditRequestTags] = None,
+        task_description: str = None,
+    ):
+        self.business_type = business_type
+        # This parameter is required.
+        self.content = content
+        self.extra_info = extra_info
+        self.model_id = model_id
+        self.output_format = output_format
+        self.tags = tags
+        self.task_description = task_description
+
+    def validate(self):
+        if self.tags:
+            for k in self.tags:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.business_type is not None:
+            result['businessType'] = self.business_type
+        if self.content is not None:
+            result['content'] = self.content
+        if self.extra_info is not None:
+            result['extraInfo'] = self.extra_info
+        if self.model_id is not None:
+            result['modelId'] = self.model_id
+        if self.output_format is not None:
+            result['outputFormat'] = self.output_format
+        result['tags'] = []
+        if self.tags is not None:
+            for k in self.tags:
+                result['tags'].append(k.to_map() if k else None)
+        if self.task_description is not None:
+            result['taskDescription'] = self.task_description
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('businessType') is not None:
+            self.business_type = m.get('businessType')
+        if m.get('content') is not None:
+            self.content = m.get('content')
+        if m.get('extraInfo') is not None:
+            self.extra_info = m.get('extraInfo')
+        if m.get('modelId') is not None:
+            self.model_id = m.get('modelId')
+        if m.get('outputFormat') is not None:
+            self.output_format = m.get('outputFormat')
+        self.tags = []
+        if m.get('tags') is not None:
+            for k in m.get('tags'):
+                temp_model = RunNetworkContentAuditRequestTags()
+                self.tags.append(temp_model.from_map(k))
+        if m.get('taskDescription') is not None:
+            self.task_description = m.get('taskDescription')
+        return self
+
+
+class RunNetworkContentAuditShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        business_type: str = None,
+        content: str = None,
+        extra_info: str = None,
+        model_id: str = None,
+        output_format: str = None,
+        tags_shrink: str = None,
+        task_description: str = None,
+    ):
+        self.business_type = business_type
+        # This parameter is required.
+        self.content = content
+        self.extra_info = extra_info
+        self.model_id = model_id
+        self.output_format = output_format
+        self.tags_shrink = tags_shrink
+        self.task_description = task_description
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.business_type is not None:
+            result['businessType'] = self.business_type
+        if self.content is not None:
+            result['content'] = self.content
+        if self.extra_info is not None:
+            result['extraInfo'] = self.extra_info
+        if self.model_id is not None:
+            result['modelId'] = self.model_id
+        if self.output_format is not None:
+            result['outputFormat'] = self.output_format
+        if self.tags_shrink is not None:
+            result['tags'] = self.tags_shrink
+        if self.task_description is not None:
+            result['taskDescription'] = self.task_description
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('businessType') is not None:
+            self.business_type = m.get('businessType')
+        if m.get('content') is not None:
+            self.content = m.get('content')
+        if m.get('extraInfo') is not None:
+            self.extra_info = m.get('extraInfo')
+        if m.get('modelId') is not None:
+            self.model_id = m.get('modelId')
+        if m.get('outputFormat') is not None:
+            self.output_format = m.get('outputFormat')
+        if m.get('tags') is not None:
+            self.tags_shrink = m.get('tags')
+        if m.get('taskDescription') is not None:
+            self.task_description = m.get('taskDescription')
+        return self
+
+
+class RunNetworkContentAuditResponseBodyHeader(TeaModel):
+    def __init__(
+        self,
+        error_code: str = None,
+        error_message: str = None,
+        event: str = None,
+        session_id: str = None,
+        task_id: str = None,
+        trace_id: str = None,
+    ):
+        self.error_code = error_code
+        self.error_message = error_message
+        self.event = event
+        self.session_id = session_id
+        self.task_id = task_id
+        self.trace_id = trace_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.error_code is not None:
+            result['errorCode'] = self.error_code
+        if self.error_message is not None:
+            result['errorMessage'] = self.error_message
+        if self.event is not None:
+            result['event'] = self.event
+        if self.session_id is not None:
+            result['sessionId'] = self.session_id
+        if self.task_id is not None:
+            result['taskId'] = self.task_id
+        if self.trace_id is not None:
+            result['traceId'] = self.trace_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('errorCode') is not None:
+            self.error_code = m.get('errorCode')
+        if m.get('errorMessage') is not None:
+            self.error_message = m.get('errorMessage')
+        if m.get('event') is not None:
+            self.event = m.get('event')
+        if m.get('sessionId') is not None:
+            self.session_id = m.get('sessionId')
+        if m.get('taskId') is not None:
+            self.task_id = m.get('taskId')
+        if m.get('traceId') is not None:
+            self.trace_id = m.get('traceId')
+        return self
+
+
+class RunNetworkContentAuditResponseBodyPayloadOutput(TeaModel):
+    def __init__(
+        self,
+        text: str = None,
+    ):
+        self.text = text
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.text is not None:
+            result['text'] = self.text
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('text') is not None:
+            self.text = m.get('text')
+        return self
+
+
+class RunNetworkContentAuditResponseBodyPayloadUsage(TeaModel):
+    def __init__(
+        self,
+        input_tokens: int = None,
+        output_tokens: int = None,
+        total_tokens: int = None,
+    ):
+        self.input_tokens = input_tokens
+        self.output_tokens = output_tokens
+        self.total_tokens = total_tokens
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.input_tokens is not None:
+            result['inputTokens'] = self.input_tokens
+        if self.output_tokens is not None:
+            result['outputTokens'] = self.output_tokens
+        if self.total_tokens is not None:
+            result['totalTokens'] = self.total_tokens
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('inputTokens') is not None:
+            self.input_tokens = m.get('inputTokens')
+        if m.get('outputTokens') is not None:
+            self.output_tokens = m.get('outputTokens')
+        if m.get('totalTokens') is not None:
+            self.total_tokens = m.get('totalTokens')
+        return self
+
+
+class RunNetworkContentAuditResponseBodyPayload(TeaModel):
+    def __init__(
+        self,
+        output: RunNetworkContentAuditResponseBodyPayloadOutput = None,
+        usage: RunNetworkContentAuditResponseBodyPayloadUsage = None,
+    ):
+        self.output = output
+        self.usage = usage
+
+    def validate(self):
+        if self.output:
+            self.output.validate()
+        if self.usage:
+            self.usage.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.output is not None:
+            result['output'] = self.output.to_map()
+        if self.usage is not None:
+            result['usage'] = self.usage.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('output') is not None:
+            temp_model = RunNetworkContentAuditResponseBodyPayloadOutput()
+            self.output = temp_model.from_map(m['output'])
+        if m.get('usage') is not None:
+            temp_model = RunNetworkContentAuditResponseBodyPayloadUsage()
+            self.usage = temp_model.from_map(m['usage'])
+        return self
+
+
+class RunNetworkContentAuditResponseBody(TeaModel):
+    def __init__(
+        self,
+        header: RunNetworkContentAuditResponseBodyHeader = None,
+        payload: RunNetworkContentAuditResponseBodyPayload = None,
+        request_id: str = None,
+    ):
+        self.header = header
+        self.payload = payload
+        # Id of the request
+        self.request_id = request_id
+
+    def validate(self):
+        if self.header:
+            self.header.validate()
+        if self.payload:
+            self.payload.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.header is not None:
+            result['header'] = self.header.to_map()
+        if self.payload is not None:
+            result['payload'] = self.payload.to_map()
+        if self.request_id is not None:
+            result['requestId'] = self.request_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('header') is not None:
+            temp_model = RunNetworkContentAuditResponseBodyHeader()
+            self.header = temp_model.from_map(m['header'])
+        if m.get('payload') is not None:
+            temp_model = RunNetworkContentAuditResponseBodyPayload()
+            self.payload = temp_model.from_map(m['payload'])
+        if m.get('requestId') is not None:
+            self.request_id = m.get('requestId')
+        return self
+
+
+class RunNetworkContentAuditResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: RunNetworkContentAuditResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = RunNetworkContentAuditResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
 class RunScriptChatRequest(TeaModel):
     def __init__(
         self,